requirements.txt
streamlit>=1.30.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
plotly>=5.18.0
Pillow>=10.0.0
